+++ src/models/inbound/thirdparty.transactions.model.ts
@@ -25,7 +25,7 @@
  --------------
  ******/
 
-import { BackendRequests } from './backend-requests'
+import { SDKRequests } from '~/shared/sdk-requests'
 import {
   Logger as SDKLogger,
   MojaloopRequests,
@@ -41,11 +41,12 @@ import { ThirdpartyTransactionStatus } from '../pispTransaction.interface'
 
 export interface InboundThridpartyTransactionsModelConfig {
   logger: SDKLogger.Logger
-  backendRequests: BackendRequests
+  sdkRequests: SDKRequests
   mojaloopRequests: MojaloopRequests
   thirdpartyRequests: ThirdpartyRequests
 }
 
+// TODO: replace this model by DFSPTransactionModel as described in docs/sequence/PISPTransactionApi.puml
 export class InboundThridpartyTransactionsModel {
   protected config: InboundThridpartyTransactionsModelConfig
 
@@ -57,14 +58,16 @@ export class InboundThridpartyTransactionsModel {
     return this.config.logger
   }
 
-  protected get backendRequests (): BackendRequests {
-    return this.config.backendRequests
+  protected get sdkRequests (): SDKRequests {
+    return this.config.sdkRequests
   }
 
   protected get mojaloopRequests (): MojaloopRequests {
     return this.config.mojaloopRequests
   }
 
+  // RequestToPay endpoint was used by POC PISPTransaction code
+  // TODO: use it when PISPMerchantTransaction flow will be implemented in future
   async requestToPayTransfer (
     inRequest: InboundThirdpartyTransactionPostRequest
   ): Promise<OutboundRequestToPayTransferPostResponse> {
@@ -98,19 +101,19 @@ export class InboundThridpartyTransactionsModel {
     }
     this.logger.push({ requestToPayTransfer }).info('requestToPayTransfer: requestToPayTransfer')
 
-    const response = await this.backendRequests.requestToPayTransfer(
+    const response = await this.sdkRequests.requestToPayTransfer(
       requestToPayTransfer
     ) as OutboundRequestToPayTransferPostResponse
 
     this.logger.push({ response }).info('requestToPayTransfer: response')
 
     // optionally notify via PATCH
-    if (config.SHARED.NOTIFY_ABOUT_TRANSFER_URI) {
+    if (config.SHARED.SDK_NOTIFY_ABOUT_TRANSFER_URI) {
       const transactionStatus: ThirdpartyTransactionStatus = {
         transactionId: inRequest.transactionRequestId,
         transactionRequestState: 'ACCEPTED'
       }
-      await this.backendRequests.notifyAboutTransfer(transactionStatus, inRequest.transactionRequestId)
+      await this.sdkRequests.notifyThirdpartyAboutTransfer(transactionStatus, inRequest.transactionRequestId)
     }
 
     return response