+++ test/integration/outbound/accounts.test.ts
@@ -27,25 +27,25 @@
 
 import axios from 'axios'
 
-describe('GET /accounts/{ID}', (): void => {
-  const scenariosURI = `http://127.0.0.1:4056/accounts/username1234`
+describe('GET /accounts/{fspId}/{userId}', (): void => {
   const expectedResp = {
     accounts: [
       {
-        "accountNickname": "dfspa.user.nickname1",
-        "id": "dfspa.username.1234",
-        "currency": "ZAR"
+        accountNickname: 'dfspa.user.nickname1',
+        id: 'dfspa.username.1234',
+        currency: 'ZAR'
       },
       {
-        "accountNickname": "dfspa.user.nickname2",
-        "id": "dfspa.username.5678",
-        "currency": "USD"
+        accountNickname: 'dfspa.user.nickname2',
+        id: 'dfspa.username.5678',
+        currency: 'USD'
       }
     ],
-    currentState: "COMPLETED"
+    currentState: 'COMPLETED'
   }
 
   it('PISP requests DFSP to return user accounts for linking', async (): Promise<void> => {
+    const scenariosURI = 'http://127.0.0.1:4006/accounts/dfspa/username1234'
     // Act
     const response = await axios.get(scenariosURI)
 
@@ -55,5 +55,24 @@ describe('GET /accounts/{ID}', (): void => {
 
     // Assert state machine state
     expect(response.data.currentState).toEqual('COMPLETED')
+  })
+
+  it('PISP requests DFSP: Expect ID not found', async (): Promise<void> => {
+    const scenariosURI = 'http://127.0.0.1:4006/accounts/dfspa/test'
+    const idNotFoundResp = {
+      accounts: [],
+      errorInformation: {
+        errorCode: '3200',
+        errorDescription: 'Generic ID not found'
+      },
+      currentState: 'COMPLETED'
+    }
+
+    await axios.get(scenariosURI)
+      .catch(error => {
+        // Assert
+        expect(error.response.status).toBe(500)
+        expect(error.response.data).toEqual(expect.objectContaining(idNotFoundResp))
+      })
   })
 })